clear code

+++ rbmp/src/value/ext/mod.rs
@@ -40,7 +40,7 @@ impl ValueExt for Value {
             Value::I32(v) => Unexpected::Signed(v as i64),
             Value::I64(v) => Unexpected::Signed(v),
             Value::U32(v) => Unexpected::Unsigned(v as u64),
-            Value::U64(v) => Unexpected::Unsigned(v as u64),
+            Value::U64(v) => Unexpected::Unsigned(v),
             Value::F32(v) => Unexpected::Float(v as f64),
             Value::F64(v) => Unexpected::Float(v),
             Value::String(ref v) => {
